+++ plone/app/event/ical/importer.py
@@ -38,7 +38,8 @@ import urllib2
 
 
 def ical_import(container, ics_resource, event_type,
-                sync_strategy=base.SYNC_KEEP_NEWER):
+                sync_strategy=base.SYNC_KEEP_NEWER,
+                workflow_trans=None):
     cal = icalendar.Calendar.from_ical(ics_resource)
     events = cal.walk('VEVENT')
 
@@ -146,6 +147,7 @@ def ical_import(container, ics_resource, event_type,
         # TODO: better use plone.api for content creation, from which some of
         # the code here is copied
 
+        content = None
         new_content_id = None
         existing_event = None
         sync_uid = _get_prop('UID', item)
@@ -175,6 +177,8 @@ def ical_import(container, ics_resource, event_type,
                                     title=title,
                                     description=description)
             content = container[new_content_id]
+
+        assert(content)  # At this point, a content must be available.
 
         event = IEventAccessor(content)
         event.title = title
@@ -211,6 +215,11 @@ def ical_import(container, ics_resource, event_type,
         else:
             transaction.savepoint(optimistic=True)
 
+        if workflow_trans:
+            # Let it crash! if a WorkflowException occurrs.
+            workflow = getToolByName(container, 'portal_workflow')
+            workflow.doActionFor(content, workflow_trans)
+
         # Do this at the end, otherwise it's overwritten
         if ext_modified:
             event.last_modified = ext_modified
@@ -230,6 +239,16 @@ class IIcalendarImportSettings(Interface):
                     u"importing icalendar resources."),
         vocabulary='plone.app.event.EventTypes',
         required=True
+    )
+
+    workflow_trans = schema.Choice(
+        title=_('ical_import_workflow_trans_title', default=u'Workflow'),
+        description=_(
+            'ical_import_workflow_trans_desc',
+            default=u"Workflow transition to apply for newly imported "
+                    u"contents."),
+        vocabulary='plone.app.vocabularies.WorkflowTransitions',
+        required=False
     )
 
     ical_url = schema.URI(
@@ -294,6 +313,7 @@ class IcalendarImportSettingsForm(form.Form):
         data['event_type'] = settings.event_type
         data['ical_url'] = settings.ical_url
         data['sync_strategy'] = settings.sync_strategy
+        data['workflow_trans'] = settings.workflow_trans
         return data
 
     def save_data(self, data):
@@ -301,6 +321,7 @@ class IcalendarImportSettingsForm(form.Form):
         settings.ical_url = data['ical_url']
         settings.event_type = data['event_type']
         settings.sync_strategy = data['sync_strategy']
+        settings.workflow_trans = data['workflow_trans']
 
     @button.buttonAndHandler(u'Save')
     def handleSave(self, action):
@@ -328,6 +349,7 @@ class IcalendarImportSettingsForm(form.Form):
         ical_url = data['ical_url']
         event_type = data['event_type']
         sync_strategy = data['sync_strategy']
+        workflow_trans = data['workflow_trans']
 
         if ical_file or ical_url:
 
@@ -343,7 +365,8 @@ class IcalendarImportSettingsForm(form.Form):
                 self.context,
                 ics_resource=ical_resource,
                 event_type=event_type,
-                sync_strategy=sync_strategy
+                sync_strategy=sync_strategy,
+                workflow_trans=workflow_trans
             )
 
             count = import_metadata['count']